CH 5: Fix the v3.9 HVC warnings about changing MU in WRF
The hybrid vertical coordinate (HVC) in WRF was introduced in
v3.9 via a large number of ifdefs. The code was fragile, as any
attempt to modify one of the 30-ish mu fields would likely
break the code.

With the v4.0 release, the ifdefs were replaced with the "in place"
1d arrays. Therefore there is no more need to warn users about the
side effects of changing one of the mu fields.